서비스 로직 추가

+++ src/main/java/com/ojt/toyproject/book/BookServiceImpl.java
@@ -2,12 +2,15 @@ package com.ojt.toyproject.book;
 
 
 
+import com.ojt.toyproject.Pagination;
 import com.ojt.toyproject.SearchDto;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
+@Slf4j
 public class BookServiceImpl implements BookService{
     private final BookMapper bookMapper;
 
@@ -23,6 +26,8 @@ public class BookServiceImpl implements BookService{
 
     @Override
     public List<BookInfoDto> getBookInfoList(SearchDto searchDto) {
+        Pagination pagination = new Pagination(searchDto);
+        searchDto.setPagination(pagination);
         List<BookInfoDto> bookInfoDtoList = bookMapper.getBookInfoList(searchDto);
         return bookInfoDtoList;
     }
@@ -32,14 +37,41 @@ public class BookServiceImpl implements BookService{
         bookMapper.updateBookInfo(bookInfoDto);
     }
 
+    //isbn으로 현재 대출중인 책의 수량 조회
+    @Override
+    public int getRentCountByIsbn(Long isbn) {
+        int rentCount = bookMapper.getBookInfoByIsbn(isbn).getRentCount();
+        return rentCount;
+    }
+
+    @Override
+    public int getAvailableCountByIsbn(Long isbn) {
+        int availableCount = bookMapper.getBookInfoByIsbn(isbn).getAvailableCount();
+        return availableCount;
+    }
+
     @Override
     public void deleteBookInfo(Long isbn) {
-        bookMapper.deleteBookInfo(isbn);
+        if (getRentCountByIsbn(isbn) == 0){
+            bookMapper.deleteBookInfo(isbn);
+        }else {
+            log.info("삭제 불가");
+        }
     }
 
     @Override
     public void deleteBookInfo(List<Long> isbnList) {
-        bookMapper.deleteBookInfos(isbnList);
+        for (int i=0; i<isbnList.size(); i++){
+            if (getRentCountByIsbn(isbnList.get(i))!=0){
+                isbnList.remove(i);
+            }
+        }
+        if (isbnList.size()!=0){
+            bookMapper.deleteBookInfos(isbnList);
+        }else {
+            log.info("삭제 가능한 도서가 없음");
+        }
+
     }
 
 
@@ -52,8 +84,8 @@ public class BookServiceImpl implements BookService{
     }
 
     @Override
-    public List<BookDto> getBookList() {
-        List<BookDto> bookDtoList = bookMapper.getBookList();
+    public List<BookDto> getBookList(SearchDto searchDto) {
+        List<BookDto> bookDtoList = bookMapper.getBookList(searchDto);
         return bookDtoList;
     }
